# smart-health-ai/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Smart Healthcare AI API",
        "description": "REST API for AI-powered Smart Healthcare Assistant System",
        "contact": {
            "email": "admin@smarthealthcare.ai"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/consultations": {
            "get": {
                "tags": [
                    "Consultations"
                ],
                "summary": "Get all consultations (paginated)",
                "operationId": "3d6bb71f050514597f61016b339b8a12",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Consultations"
                ],
                "summary": "Send a consultation message (triggers AI triage)",
                "operationId": "691cf6ae12418e03b55f0718dd9641ee",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "message"
                                ],
                                "properties": {
                                    "message": {
                                        "type": "string",
                                        "example": "I have a fever and sore throat"
                                    },
                                    "patient_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "session_id": {
                                        "type": "string",
                                        "example": "uuid-session-id"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Consultation created with AI response"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/consultations/{id}": {
            "get": {
                "tags": [
                    "Consultations"
                ],
                "summary": "Get a specific consultation",
                "operationId": "8bea096fabf6548f23bf7cf7141142d6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Consultations",
            "description": "AI-powered consultation endpoints"
        }
    ]
}